# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current authenticated user",
        "tags": ["users"]
      }
    },
    "/users/me/locale": {
      "patch": {
        "operationId": "UsersController_updateLocalePreference",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocaleDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Locale updated"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the current user locale preference",
        "tags": ["users"]
      }
    },
...
  },
  "info": {
    "title": "Power Budget API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateLocaleDto": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "enum": ["en", "uk", "ru", "pl"],
            "example": "en"
          }
        },
        "required": ["locale"]
      },
...
    }
  }
}